Move clear function to page switch
